Divide classes into individual files
